test_requirements.txt
coverage>=3.7.1
flake8>=1.5.0
mock>=1.0.1
numpydoc>=0.6.0
nose>=1.0.0
pep8==1.7.0
pytest==2.9.2
pylint>=1.5.4
